Commit made by the Bioconductor Git-SVN bridge.
Consists of 1 commit.

Commit information:

SVN Revision number: 97070
Commit message:
Add some config options to track the version of R and bioc used
by the workflow and single package builders, so that this information
does not need to be hardcoded elsewhere (where we forget to change it
with deleterious consequences). Now we just have to remember to change it in
one place.
Also, copy config.yaml to the assets directory with every site build, this
enables it to be accessible at /config.yaml so that various build tools
can more easily access the yaml config.

Committed by d.tenenbaum
Committed at: 2014-11-24T18:21:14.846566Z