Use redis to keep track of build status and decide whether rss feeds should be updated. (Also add redis to list of gems needed for website.)
The rule now is (I think) a little different than before:

If redis does not know about the package at all, then make an rss file.
If the current build status is different than what is in redis,
update the rss file. If build status has not changed, do not update
the rss file.

This should result in notification whenever the build
status changes. So you should be notified when an error happens and
when it is resolved but not every day as long as things remain the same.